Do symbol lookups internally
I'm considering to deprecate `NonContiguousCategoricalEncoderModel` and
`NonContiguousCategoricalDecoderModel` in `constriction` to simplify the
API and decouple entropy coding from lookups in symbol tables. This
means that applications have to use `ContiguousCategoricalEntropyModel`
instead and do the conversion between indices and symbols themselves.
This commit already implements this update so that we can test its
impact on performance.